pyproject.toml
[tool.poetry]
name = "cs5246-text-mining-proj"
version = "0.1.0"
description = ""
authors = ["Shivansh Bakshi <work@shivanshbakshi.dev>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.2.3"
tqdm = "^4.67.1"
requests = "^2.32.3"
newspaper3k = "^0.2.8"
gdeltdoc = "^1.12.0"
matplotlib = "^3.10.1"
googlemaps = "^4.10.0"
pycountry = "^24.6.1"
scikit-learn = "^1.6.1"
nltk = "^3.9.1"
contractions = "^0.1.73"
sentence-transformers = "^4.1.0"
spacy = "^3.8.5"
accelerate = "^1.6.0"
regex = "^2024.11.6"
seaborn = "^0.13.2"
folium = "^0.19.5"
haversine = "^2.9.0"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.27.1"
gensim = "^4.3.3"
geonamescache = "^2.0.0"
dateparser = "^1.2.1"
vadersentiment = "^3.3.2"
httpx = "^0.28.1"
langchain = "^0.3.23"
langchain-google-genai = "^2.1.3"
python-dotenv = "^1.1.0"
tiktoken = "^0.9.0"
ipywidgets = "^8.1.6"


[tool.poetry.group.dev.dependencies]
pandas-stubs = "^2.2.3.250308"
types-tqdm = "^4.67.0.20250404"
black = "^25.1.0"
types-dateparser = "^1.2.0.20250408"
types-regex = "^2024.11.6.20250403"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
